+++ apps/nextjs/src/components/AppComponents/Chat/sidebar-items.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { OakBox, OakFlex, OakP } from "@oaknational/oak-components";
 import { format, isToday, isThisWeek, isThisMonth } from "date-fns";
-import { AnimatePresence, motion } from "framer-motion";
+import { AnimatePresence } from "framer-motion";
 
-import { SidebarItem } from "@/components/AppComponents/Chat/sidebar-item";
 import type { SideBarChatItem } from "@/lib/types";
+
+import RenderChats from "./render-chat-history";
 
 interface SidebarItemsProps {
   chats: SideBarChatItem[];
@@ -46,32 +46,13 @@ export function SidebarItems({ chats }: Readonly<SidebarItemsProps>) {
     return acc;
   }, {});
 
-  const renderChats = (title: string, chats: SideBarChatItem[]) => (
-    <OakBox $ml="space-between-s">
-      <OakP $font="body-2-bold" $mb="space-between-s">
-        {title}
-      </OakP>
-      <OakFlex
-        $mb="space-between-s"
-        $flexDirection="column"
-        $gap="all-spacing-2"
-      >
-        {chats.map((chat) => (
-          <motion.div key={chat.id} exit={{ opacity: 0, height: 0 }}>
-            <SidebarItem chat={chat} />
-          </motion.div>
-        ))}
-      </OakFlex>
-    </OakBox>
-  );
-
   return (
     <AnimatePresence>
-      {todayChats.length > 0 && renderChats("Today", todayChats)}
-      {lastWeekChats.length > 0 && renderChats("Last week", lastWeekChats)}
-      {lastMonthChats.length > 0 && renderChats("Last 30 days", lastMonthChats)}
+      {todayChats.length > 0 && RenderChats("Today", todayChats)}
+      {lastWeekChats.length > 0 && RenderChats("Last week", lastWeekChats)}
+      {lastMonthChats.length > 0 && RenderChats("Last 30 days", lastMonthChats)}
       {Object.entries(olderChatsByMonth).map(([monthYear, chats]) =>
-        renderChats(monthYear, chats),
+        RenderChats(monthYear, chats),
       )}
     </AnimatePresence>
   );
