+++ apps/nextjs/src/utils/convertTitleCaseToSentenceCase.ts
@@ -1,0 +1,4 @@
+export function convertTitleCaseToSentenceCase(titleCase: string) {
+  const lowerCaseTitle = titleCase.toLowerCase();
+  return lowerCaseTitle.charAt(0).toUpperCase() + lowerCaseTitle.slice(1);
+}
